Resolve init errors and clean up build leftovers (#61)
* Remove tests that will always evaluate to True
Change update backup path to mountpoint guranteed to be on the same partition

* Add checks to close #56

* Clone `log2ram` to `/tmp` to close #60

* Update initramfs and md5 hash

* Refactor `finalize_update` and `post_update` to `*-update` for consistency with other services
Fix newline error in `neon_env.conf`

* Update base image

+++ initramfs_build/overlay/scripts/check_update.sh
@@ -48,10 +48,10 @@ mv "${UPDATE_FILE}" "${ROOT_FILE}" && echo "Applied updated rootfs from ${UPDATE
 mv "${OVERLAY_PATH}" "${BACKUP_PATH}" && echo "Backed up overlay ${OVERLAY_PATH} to ${BACKUP_PATH}"
 
 # Ensure paths exist for migrated data
-[ -d "${OVERLAY_PATH}/etc/NetworkManager" ] || mkdir -p "${OVERLAY_PATH}/etc/NetworkManager"
-[ -d "${OVERLAY_PATH}/etc/ssh" ] || mkdir -p "${OVERLAY_PATH}/etc/ssh"
-[ -d "${OVERLAY_PATH}/opt/neon" ] || mkdir -p "${OVERLAY_PATH}/opt/neon"
-[ -d "${OVERLAY_PATH}/var" ] || mkdir -p "${OVERLAY_PATH}/var"
+mkdir -p "${OVERLAY_PATH}/etc/NetworkManager"
+mkdir -p "${OVERLAY_PATH}/etc/ssh"
+mkdir -p "${OVERLAY_PATH}/opt/neon"
+mkdir -p "${OVERLAY_PATH}/var"
 
 # Migrate specific data back
 rm -rf "${BACKUP_PATH}/home/neon/venv" && echo "Removed old venv"
@@ -64,9 +64,9 @@ mv "${BACKUP_PATH}/home" "${OVERLAY_PATH}/" && echo "Restored /home"
 mv "${BACKUP_PATH}/root" "${OVERLAY_PATH}/" && echo "Restored /root"
 
 # Restore signal files
-mv "${BACKUP_PATH}/opt/neon/firstboot" "${OVERLAY_PATH}/opt/neon/" && echo "Restored firstboot flag"
-mv "${BACKUP_PATH}/opt/neon/do_sj201_config" "${OVERLAY_PATH}/opt/neon/" && echo "Restored do_sj201_config flag"
-mv "${BACKUP_PATH}/opt/neon/do_bluetooth_config" "${OVERLAY_PATH}/opt/neon/" && echo "Restored do_bluetooth_config flag"
+[ -f "${BACKUP_PATH}/opt/neon/firstboot" ] && mv "${BACKUP_PATH}/opt/neon/firstboot" "${OVERLAY_PATH}/opt/neon/" && echo "Restored firstboot flag"
+[ -f "${BACKUP_PATH}/opt/neon/do_sj201_config" ] && mv "${BACKUP_PATH}/opt/neon/do_sj201_config" "${OVERLAY_PATH}/opt/neon/" && echo "Restored do_sj201_config flag"
+[ -f "${BACKUP_PATH}/opt/neon/do_bluetooth_config" ] && mv "${BACKUP_PATH}/opt/neon/do_bluetooth_config" "${OVERLAY_PATH}/opt/neon/" && echo "Restored do_bluetooth_config flag"
 
 # Restore specific `/var` paths (https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch05.html)
 [ -d "${BACKUP_PATH}/var/games" ] && mv "${BACKUP_PATH}/var/games" "${OVERLAY_PATH}/var/" && echo "Restored /var/games"
